move submit logic to hook
